Fix nhiều bug

+++ BA/QLBaiDoXe/QLBaiDoXe/DSThe.xaml.cs
@@ -1,17 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 using QLBaiDoXe.DBClasses;
 
 namespace QLBaiDoXe
@@ -31,16 +19,24 @@ namespace QLBaiDoXe
         {
             ThemThe add = new ThemThe();
             add.ShowDialog();
+            ListThe.ItemsSource = Cards.GetAllParkingCards();
         }
 
         private void Button_Click_1(object sender, RoutedEventArgs e)
         {
+            if (ListThe.Items.Count == 0)
+            {
+                MessageBox.Show("Danh sách thẻ rỗng!");
+                return;
+            }    
             if (ListThe.SelectedItems == null)
             {
                 MessageBox.Show("Hãy chọn thẻ cần xóa!");
             }
             else
             {
+                if (MessageBox.Show("Bạn có muốn xóa thẻ đã chọn?", "Xác nhận", MessageBoxButton.YesNo) == MessageBoxResult.No)
+                    return;
                 var selectedItems = (dynamic)ListThe.SelectedItems[0];
                 Cards.DeleteCard((long)selectedItems.ParkingCardID);
                 MessageBox.Show("Đã xóa thẻ thành công!");
